pyproject.toml
[tool.poetry]
name = "credit"
version = "0.1.0"
description = ""
authors = ["Timur Vakhitov <t.vakhitov@astondevs.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
black = "^23.1.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.5"
djangorestframework-simplejwt = "^5.2.2"
pre-commit = "^3.1.1"
pytest = "^7.2.2"
pytest-django = "^4.5.2"
drf-yasg = "^1.21.5"
pytest-snapshot = "^0.9.0"
factory-boy = "^3.2.1"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
